+++ chat2rag/api/v1/prompt.py
@@ -6,7 +6,8 @@ from sqlalchemy.orm import Session
 from chat2rag.api.schema import Error, Success
 from chat2rag.api.schema.prompt import PromptCreate
 from chat2rag.config import CONFIG
-from chat2rag.core.database import Prompt, get_db
+from chat2rag.database.connection import get_db
+from chat2rag.database.models import Prompt
 from chat2rag.logger import logger
 
 router = APIRouter()